+++ pkg/metafile/metafile.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"path/filepath"
 	"time"
 
 	"github.com/paulschiretz/pgl-backup/pkg/util"
@@ -30,8 +29,8 @@ type MetafileContent struct {
 }
 
 // Write creates and writes the .pgl-backup.meta.json file into a given directory.
-func Write(dirPath string, content *MetafileContent) error {
-	metaFilePath := filepath.Join(dirPath, MetaFileName)
+func Write(absDirPath string, content *MetafileContent) error {
+	metaFilePath := util.DenormalizedAbsPath(absDirPath, MetaFileName)
 	jsonData, err := json.MarshalIndent(content, "", "  ")
 	if err != nil {
 		return fmt.Errorf("could not marshal meta data: %w", err)
@@ -50,8 +49,8 @@ func Write(dirPath string, content *MetafileContent) error {
 
 // Read opens and parses the .pgl-backup.meta.json file in a given directory.
 // It returns the parsed metadata or an error if the file cannot be read or parsed.
-func Read(dirPath string) (MetafileContent, error) {
-	metaFilePath := filepath.Join(dirPath, MetaFileName)
+func Read(absDirPath string) (MetafileContent, error) {
+	metaFilePath := util.DenormalizedAbsPath(absDirPath, MetaFileName)
 	metaFile, err := os.Open(metaFilePath)
 	if err != nil {
 		// Note: os.IsNotExist errors are handled by the caller.